+++ camcomskin/padova/browser/tile_collection_renderers.py
@@ -21,51 +21,46 @@ class HelpersView(BrowserView):
     A set of helper functions for tile collection views.
     """
 
-    def get_image_tag(self,
-                      item,
-                      scale='thumb',
-                      direction='down',
-                      css_class=''):
+    def get_image_tag(self, item, scale="thumb", direction="down", css_class=""):
         try:
             scale_view = api.content.get_view(
-                name='images',
-                context=item,
-                request=self.request,
+                name="images", context=item, request=self.request
             )
-            return scale_view.scale(
-                'image', scale=scale, direction=direction
-                ).tag(css_class=css_class)
+            return scale_view.scale("image", scale=scale, direction=direction).tag(
+                css_class=css_class
+            )
         except (InvalidParameterError, POSKeyError, AttributeError, KeyError):
             # The object doesn't have an image field
             return ""
 
-    def get_bg_url(self, item, scale='thumb'):
+    def get_bg_url(self, item, scale="thumb"):
         try:
             scale_view = api.content.get_view(
-                name='images',
-                context=item,
-                request=self.request,
+                name="images", context=item, request=self.request
             )
-            return 'background-image: url("' +\
-                   str(scale_view.scale('image', scale=scale).url) + '");'
+            return (
+                'background-image: url("'
+                + str(scale_view.scale("image", scale=scale).url)
+                + '");'
+            )
         except (InvalidParameterError, POSKeyError, AttributeError):
             # The object doesn't have an image field
             return ""
 
     def get_formatted_date(self, item):
-        '''
+        """
         return a formatted date
-        '''
+        """
         effective = item.effective
         if effective.year() == 1969:
             # not yet published
             return {}
         return {
-            'weekday': u'weekday_{0}'.format(effective.aDay().lower()),
-            'month': u'month_{0}'.format(effective.aMonth().lower()),
-            'month_abbr': u'month_{0}_abbr'.format(effective.aMonth().lower()),
-            'day': effective.day(),
-            'year': effective.year()
+            "weekday": u"weekday_{0}".format(effective.aDay().lower()),
+            "month": u"month_{0}".format(effective.aMonth().lower()),
+            "month_abbr": u"month_{0}_abbr".format(effective.aMonth().lower()),
+            "day": effective.day(),
+            "year": effective.year(),
         }
 
     def get_event_dates(self, event):
@@ -74,38 +69,44 @@ class HelpersView(BrowserView):
         """
 
         months = {
-            '1': 'Gennaio',
-            '2': 'Febbraio',
-            '3': 'Marzo',
-            '4': 'Aprile',
-            '5': 'Maggio',
-            '6': 'Giugno',
-            '7': 'Luglio',
-            '8': 'Agosto',
-            '9': 'Settembre',
-            '10': 'Ottobre',
-            '11': 'Novembre',
-            '12': 'Dicembre'
+            "1": "Gennaio",
+            "2": "Febbraio",
+            "3": "Marzo",
+            "4": "Aprile",
+            "5": "Maggio",
+            "6": "Giugno",
+            "7": "Luglio",
+            "8": "Agosto",
+            "9": "Settembre",
+            "10": "Ottobre",
+            "11": "Novembre",
+            "12": "Dicembre",
         }
 
         date_string = ""
         start_date_parts = [
-            event.startDate.day(), months[str(event.startDate.month())],
-            event.startDate.year()
+            event.startDate.day(),
+            months[str(event.startDate.month())],
+            event.startDate.year(),
         ]
-        startDate = ' '.join([str(x) for x in start_date_parts])
+        startDate = " ".join([str(x) for x in start_date_parts])
 
         if event.startDate.Date() == event.endDate.Date():
-            date_string = startDate + ", ore " +\
-                event.startDate.TimeMinutes() + " - " +\
-                event.endDate.TimeMinutes()
+            date_string = (
+                startDate
+                + ", ore "
+                + event.startDate.TimeMinutes()
+                + " - "
+                + event.endDate.TimeMinutes()
+            )
 
         else:
             end_date_parts = [
-                event.endDate.day(), months[str(event.endDate.month())],
-                event.endDate.year()
+                event.endDate.day(),
+                months[str(event.endDate.month())],
+                event.endDate.year(),
             ]
-            endDate = ' '.join([str(x) for x in end_date_parts])
+            endDate = " ".join([str(x) for x in end_date_parts])
             date_string = "dal " + startDate + " al " + endDate
 
         return date_string
@@ -192,6 +193,7 @@ class SightsView(BrowserView):
     """
     Custom view that shows sights
     """
+
     implements(ICollectionTileRenderer)
 
     display_name = _("Sights layout")
@@ -201,6 +203,7 @@ class NewsHighlightView(BrowserView):
     """
     Custom view that shows an highlighted news
     """
+
     implements(ICollectionTileRenderer)
 
     display_name = _("News highlight")
@@ -228,6 +231,7 @@ class NewsBigPhotoView(BrowserView):
     """
     Custom view that shows a news with a big photo on the background
     """
+
     implements(ICollectionTileRenderer)
 
     display_name = _("News with big photo")
@@ -237,6 +241,7 @@ class NewsAreaTematicaView(BrowserView):
     """
     Custom view that shows news in area tematica
     """
+
     implements(ICollectionTileRenderer)
 
     display_name = _("News in area tematica")
@@ -246,6 +251,7 @@ class ServiziAreaTematicaView(BrowserView):
     """
     Custom view that shows servizi in area tematica
     """
+
     implements(ICollectionTileRenderer)
 
     display_name = _("Servizi in area tematica")
@@ -291,3 +297,9 @@ class OnlineServicesView(BrowserView):
     implements(ICollectionTileRenderer)
 
     display_name = _("Layout servizi online")
+
+
+class CCIAACollectionTileView(BrowserView):
+    implements(ICollectionTileRenderer)
+
+    display_name = _("Layout collezione CCIAA")